% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aglm_output_summary.r
\name{aglm_output_summary}
\alias{aglm_output_summary}
\title{Produce data summary of aglm model}
\usage{
aglm_output_summary(
  model,
  model_data,
  summary_fields,
  penalties = c(model@lambda.1se, model@lambda.min, 0)
)
}
\arguments{
\item{model}{aglm or cv.aglm object}

\item{model_data}{dataframe that produced the aglm object}

\item{summary_fields}{Character vector of field names to sum}

\item{penalties}{List of penalty levels for which to produce coefficients.  Equivalent to s in \link[aglm]{predict}.  Defaults to
showing coefficients for \code{lambda.1se}, \code{lambda.min}, and 0 penalty.}
}
\value{
Dataframe containing all modeled variables and levels from the aglm object and the associated summarized variables from the model dataset.
}
\description{
Produce a dataframe containing summary of selected fields given an aglm model, the associated dataframe that produced it, and a character list of columns to summarize.
Applies the model's binning processes to the source data to facilitate one to one comparisons of model coefficients and raw data.
}
\examples{
data('Boston', package = 'MASS')
model = cv.aglm(y = Boston$medv, x = Boston \%>\% select(-medv))
data_summary = aglm_output_summary(model, Boston, summary_fields = c('crim','age'))
}
