% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_by_append.r
\name{join_by_append}
\alias{join_by_append}
\title{Join-by append}
\usage{
join_by_append(jb1, jb2)
}
\arguments{
\item{jb1}{Dplyr join_by specification}

\item{jb2}{Dplyr join_by specification}
}
\value{
Dplyr join_by specification combining all specifications of jb1 and jb2.
}
\description{
Function to append new conditions to a dplyr "join_by" specification.  Useful when multiple related join specifications are used on the same dataframe to avoid writing each specification out multiple times.
}
\examples{
join_cond_1 = join_by(effective_date <= evaluation_period_end)
join_cond_2 = join_by(effective_date >= evaluation_period_start)
join_cond_combined = join_by_append(join_cond_1, join_cond_2)
}
