% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_levels.r
\name{order_levels}
\alias{order_levels}
\title{Re-order levels}
\usage{
order_levels(x)
}
\arguments{
\item{x}{Character vector}
}
\value{
Character vector of unique levels, ordered by the numeric portion of the string detected
}
\description{
Logically order levels of mixed text/numeric strings based on the numbers in the string
Output is only the raw unique levels ordered in this fashion.  Factors will need to be created manually.
Levels without any numeric part are ordered randomly.
}
\examples{
sample_vector = sapply(0:10, function(x) paste(c(sample(LETTERS,5,replace=TRUE), sample(0:9,4,replace=TRUE)), collapse = ''))
order_levels(sample_vector)
}
